fix: favoriteBtn.ts import export Eliminated.
