--- src/common/index.ts
@@ -1,1 +1,0 @@
-export const LOCALHOST_URL = 'http://localhost:3000';
